third_party/blink/web_tests/external/wpt/fetch/metadata/generated/element-link-prefetch.https.optional.sub-expected.txt

This is a testharness.js-based test.
[FAIL] sec-fetch-dest attributes: as=audio
  assert_array_equals: expected property 0 to be "audio" but got "empty" (expected array ["audio"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=document
  assert_array_equals: expected property 0 to be "document" but got "empty" (expected array ["document"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=embed
  assert_array_equals: expected property 0 to be "embed" but got "empty" (expected array ["embed"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=fetch
  assert_array_equals: expected property 0 to be "fetch" but got "empty" (expected array ["fetch"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=font
  assert_array_equals: expected property 0 to be "font" but got "empty" (expected array ["font"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=image
  assert_array_equals: expected property 0 to be "image" but got "empty" (expected array ["image"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=object
  assert_array_equals: expected property 0 to be "object" but got "empty" (expected array ["object"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=script
  assert_array_equals: expected property 0 to be "script" but got "empty" (expected array ["script"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=style
  assert_array_equals: expected property 0 to be "style" but got "empty" (expected array ["style"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=track
  assert_array_equals: expected property 0 to be "track" but got "empty" (expected array ["track"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=video
  assert_array_equals: expected property 0 to be "video" but got "empty" (expected array ["video"] got ["empty"])
[FAIL] sec-fetch-dest attributes: as=worker
  assert_array_equals: expected property 0 to be "worker" but got "empty" (expected array ["worker"] got ["empty"])
Harness: the test ran to completion.
